+++ c/column_bool.cc
@@ -16,8 +16,8 @@ SType BoolColumn::stype() const {
   return SType::BOOL;
 }
 
-py::oobj BoolColumn::get_value_at_index(int64_t i) const {
-  int64_t j = ri.nth(i);
+py::oobj BoolColumn::get_value_at_index(size_t i) const {
+  size_t j = ri.nth(i);
   int8_t x = elements_r()[j];
   return ISNA(x)? py::None() : x? py::True() : py::False();
 }
