+++ c/column_real.cc
@@ -20,8 +20,8 @@ SType RealColumn<T>::stype() const {
 }
 
 template <typename T>
-py::oobj RealColumn<T>::get_value_at_index(int64_t i) const {
-  int64_t j = (this->ri).nth(i);
+py::oobj RealColumn<T>::get_value_at_index(size_t i) const {
+  size_t j = (this->ri).nth(i);
   T x = this->elements_r()[j];
   return ISNA<T>(x)? py::None() : py::ofloat(x);
 }
